Issue #42: Refactor basket tests
Break basket tests out from a single, humongous file into a bunch of
files, none over 1000 lines long and most much shorter than that.
