+++ news/tests/test_models.py
@@ -1,0 +1,28 @@
+from django.test import TestCase
+
+from news import models
+
+
+class SubscriberTest(TestCase):
+    def test_get_and_sync_creates(self):
+        """
+        Subscriber.objects.get_and_sync() should create if email doesn't exist.
+        """
+        with self.assertRaises(models.Subscriber.DoesNotExist):
+            models.Subscriber.objects.get(email='dude@example.com')
+
+        models.Subscriber.objects.get_and_sync('dude@example.com', 'asdfjkl')
+        sub = models.Subscriber.objects.get(email='dude@example.com')
+        self.assertEqual(sub.token, 'asdfjkl')
+
+    def test_get_and_sync_updates(self):
+        """
+        Subscriber.objects.get_and_sync() should update token if it doesn't
+        match.
+        """
+        models.Subscriber.objects.create(email='dude@example.com',
+                                         token='asdf')
+
+        models.Subscriber.objects.get_and_sync('dude@example.com', 'asdfjkl')
+        sub = models.Subscriber.objects.get(email='dude@example.com')
+        self.assertEqual(sub.token, 'asdfjkl')
